Implemented controller via Unit of Work pattern
Added the TodoListUow for application level grouping of operations. This
allowed me to put the logic the controller needed into one place that
can then be tested via mocking the TodoItemRepository.

The controller had the direct calls to EF removed and replaced with
calls to TodoListUow
NSubstitute was used to mock the repository in the unit tests.

I removed Details & Delete as the requirements didn't need these.